refactor: Create vertical grid component module.
