+++ app/src/main/java/com/guru/composecookbook/ui/demoapps/spotify/SpotifyHome.kt
@@ -28,7 +28,7 @@ import com.guru.composecookbook.R
 import com.guru.composecookbook.theme.graySurface
 import com.guru.composecookbook.theme.typography
 import com.guru.composecookbook.ui.demoapps.spotify.data.SpotifyDataProvider
-import com.guru.composecookbook.ui.utils.VerticalGrid
+import com.guru.composecookbook.verticalgrid.VerticalGrid
 import com.guru.composecookbook.theme.modifiers.horizontalGradientBackground
 
 @Composable
