+++ app/src/main/java/com/guru/composecookbook/ui/demoapps/tiktok/profile/TikTokProfile.kt
@@ -26,7 +26,7 @@ import com.guru.composecookbook.theme.ComposeCookBookTheme
 import com.guru.composecookbook.theme.tiktokRed
 import com.guru.composecookbook.ui.demoapps.spotify.data.Album
 import com.guru.composecookbook.ui.demoapps.spotify.data.SpotifyDataProvider
-import com.guru.composecookbook.ui.utils.VerticalGrid
+import com.guru.composecookbook.verticalgrid.VerticalGrid
 
 @Composable
 fun TikTokProfile(userId: String = "10", navHostController: NavHostController) {
