+++ components/verticalgrid/build.gradle.kts
@@ -1,0 +1,12 @@
+import com.guru.composecookbook.build.dependencies.*
+
+plugins {
+    /**
+     * See [common-compose-module-configs-script-plugin.gradle.kts] file
+     */
+    id("common-compose-module-configs-script-plugin")
+}
+
+dependencies {
+    addComposeOfficialDependencies()
+}
