+++ settings.gradle.kts
@@ -10,3 +10,4 @@ include(":demos:moviesapp:data")
 include(":components:charts")
 include(":components:tags")
 include(":components:carousel")
+include(":components:verticalgrid")
